Add files via upload
My first post, I hope all is well. This is a very basic IDE I wrote with the help of Grok and DeepSeek. If it has potential, I hope someone more apt with coding (Python) might help it  (perhaps a lot of help) fly higher. Else, we will just let it RIP. Thank you for taking a look. My special thanks to Anobium for inviting me to post it here.